LPD-56673 Remove loading/setLoading logic from MultipleSelection. Because no specific logic was left in MultipleSelection, I have moved handleAsyncOptions to base component and called it directly.
